Refactor imports to move ChannelType to libs.types for improved organization and clarity
