+++ integrations/standard_io/parser.py
@@ -7,8 +7,7 @@ from typing import Any, cast
 
 from integrations.base.interface import MessageParserDataMixin, MessageParserInterface
 from integrations.protocols import MsgData, PostData, StatusData
-from libs.domain.datamodels import ChannelType
-from libs.types import MessageStatus
+from libs.types import ChannelType, MessageStatus
 
 
 class MessageParser(MessageParserDataMixin, MessageParserInterface):
