+++ libs/data/lookup.py
@@ -10,8 +10,9 @@ from typing import TYPE_CHECKING, Any, Optional, Union, cast
 
 import libs.global_value as g
 from libs.data import loader
-from libs.domain.datamodels import ChannelType, CommandType
+from libs.domain.datamodels import CommandType
 from libs.domain.score import GameResult
+from libs.types import ChannelType
 from libs.utils import dbutil
 from libs.utils.timekit import ExtendedDatetime as ExtDt
 from libs.utils.timekit import Format
