+++ libs/domain/datamodels.py
@@ -44,23 +44,6 @@ class CommandType(StrEnum):
     """未定義"""
 
 
-class ChannelType(StrEnum):
-    """チャンネルタイプ"""
-
-    CHANNEL = "normal"
-    """通常チャンネル"""
-    PRIVATE = "private"
-    """プライベートチャンネル"""
-    DIRECT_MESSAGE = "direct_message"
-    """ダイレクトメッセージ"""
-    HOME_APP = "home_app"
-    """Slackのホームアプリ"""
-    SEARCH = "search_api"
-    """検索API"""
-    UNDETERMINED = "undetermined"
-    """未定義状態"""
-
-
 @dataclass
 class GameInfo:
     """ゲーム集計情報"""
